Task 211: adapt TreeView for the case where a version is linked
or when a version is the linker. When a version id is given then
also follow the links for the config item id. When a config item id
is given then also follow the links for the most recent version.
This means that in the entry point only the version id of the origin is needed.

Use more specific IDs for the edges. The ID 'C1V2' indicates config item id 1 and version id 2.

Also a bit of tidying. Shorten ConfigItemVersionID to VersionID as
this seems to be the common usage. Try to use less convoluted data structures.

Eliminate trailing semicolon in data passed to the JS tree library.

Adapt Core.Agent.TreeVies.js to check early for empty strings.