Issue #54: Simple fix preventing empty search.
An empty ConfigItemSearch returns all ConfigItems (default Limit 10000).
In case that exactly 1 config item exists on the system, it returns this
one, resulting in it being updated instead of a new one being created.

The core issue is that, if the identifier attribute is not present in
the request payload, an empty search is done currently.